Revert "IMS-VT: Modification of data usage aidl"
This reverts commit 3d008117add89bb421f9b5e6610c07cec407a3cf.

+++ telecomm/java/android/telecom/InCallService.java
@@ -372,7 +372,7 @@ public abstract class InCallService extends Service {
              *
              * @param dataUsage The updated data usage.
              */
-            public abstract void onCallDataUsageChanged(long dataUsage);
+            public abstract void onCallDataUsageChanged(int dataUsage);
 
             /**
              * Handles a change in camera capabilities.
